P4ADEV-2134 split CreateReceiptServiceImpl on class UpdatePaidDebtPositionService
