+++ src/main/java/it/gov/pagopa/pu/debtpositions/service/create/receipt/CreateReceiptServiceImpl.java
@@ -1,26 +1,14 @@
 package it.gov.pagopa.pu.debtpositions.service.create.receipt;
 
-import it.gov.pagopa.pu.debtpositions.connector.organization.service.BrokerService;
-import it.gov.pagopa.pu.debtpositions.connector.organization.service.OrganizationService;
 import it.gov.pagopa.pu.debtpositions.dto.Receipt;
-import it.gov.pagopa.pu.debtpositions.dto.generated.DebtPositionDTO;
-import it.gov.pagopa.pu.debtpositions.dto.generated.DebtPositionStatus;
 import it.gov.pagopa.pu.debtpositions.dto.generated.ReceiptDTO;
 import it.gov.pagopa.pu.debtpositions.dto.generated.ReceiptWithAdditionalNodeDataDTO;
-import it.gov.pagopa.pu.debtpositions.mapper.DebtPositionMapper;
 import it.gov.pagopa.pu.debtpositions.mapper.ReceiptMapper;
-import it.gov.pagopa.pu.debtpositions.model.DebtPosition;
-import it.gov.pagopa.pu.debtpositions.model.InstallmentNoPII;
 import it.gov.pagopa.pu.debtpositions.model.ReceiptNoPII;
 import it.gov.pagopa.pu.debtpositions.repository.ReceiptNoPIIRepository;
 import it.gov.pagopa.pu.debtpositions.repository.ReceiptPIIRepository;
-import it.gov.pagopa.pu.debtpositions.service.sync.DebtPositionSyncService;
-import it.gov.pagopa.pu.organization.dto.generated.Broker;
-import it.gov.pagopa.pu.workflowhub.dto.generated.PaymentEventType;
-import it.gov.pagopa.pu.workflowhub.dto.generated.WorkflowCreatedDTO;
 import jakarta.transaction.Transactional;
 import lombok.extern.slf4j.Slf4j;
-import org.apache.commons.lang3.tuple.Pair;
 import org.springframework.stereotype.Service;
 
 import java.util.Optional;
@@ -33,30 +21,17 @@ public class CreateReceiptServiceImpl implements CreateReceiptService {
   private final ReceiptNoPIIRepository receiptNoPIIRepository;
   private final ReceiptPIIRepository receiptPIIRepository;
   private final ReceiptMapper receiptMapper;
-  private final OrganizationService organizationService;
-  private final BrokerService brokerService;
-  private final DebtPositionSyncService debtPositionSyncService;
-  private final DebtPositionMapper debtPositionMapper;
-  private final PrimaryOrgInstallmentPaidVerifierService primaryOrgInstallmentPaidVerifierService;
-  private final InstallmentUpdateService installmentUpdateService;
+  private final UpdatePaidDebtPositionService updateInstallmentStatusOfDebtPosition;
   private final CreatePaidTechnicalDebtPositionsService createPaidTechnicalDebtPositionsService;
 
-  public CreateReceiptServiceImpl(ReceiptNoPIIRepository receiptNoPIIRepository, ReceiptPIIRepository receiptPIIRepository,
-                                  ReceiptMapper receiptMapper, OrganizationService organizationService,
-                                  BrokerService brokerService, DebtPositionSyncService debtPositionSyncService,
-                                  DebtPositionMapper debtPositionMapper, PrimaryOrgInstallmentPaidVerifierService primaryOrgInstallmentPaidVerifierService,
-                                  InstallmentUpdateService installmentUpdateService, CreatePaidTechnicalDebtPositionsService createPaidTechnicalDebtPositionsService) {
+  public CreateReceiptServiceImpl(ReceiptNoPIIRepository receiptNoPIIRepository, ReceiptPIIRepository receiptPIIRepository, ReceiptMapper receiptMapper, UpdatePaidDebtPositionService updateInstallmentStatusOfDebtPosition, CreatePaidTechnicalDebtPositionsService createPaidTechnicalDebtPositionsService) {
     this.receiptNoPIIRepository = receiptNoPIIRepository;
     this.receiptPIIRepository = receiptPIIRepository;
     this.receiptMapper = receiptMapper;
-    this.organizationService = organizationService;
-    this.brokerService = brokerService;
-    this.debtPositionSyncService = debtPositionSyncService;
-    this.debtPositionMapper = debtPositionMapper;
-    this.primaryOrgInstallmentPaidVerifierService = primaryOrgInstallmentPaidVerifierService;
-    this.installmentUpdateService = installmentUpdateService;
+    this.updateInstallmentStatusOfDebtPosition = updateInstallmentStatusOfDebtPosition;
     this.createPaidTechnicalDebtPositionsService = createPaidTechnicalDebtPositionsService;
   }
+
 
   @Override
   @Transactional
@@ -71,27 +46,11 @@ public class CreateReceiptServiceImpl implements CreateReceiptService {
     //persist receipt
     saveReceipt(receiptDTO);
 
-    //check if organization who handle the notice is managed by PU
-    boolean[] primaryOrgFound = new boolean[]{false};
-    organizationService.getOrganizationByFiscalCode(receiptDTO.getOrgFiscalCode(), accessToken)
-      .ifPresent(primaryOrg -> {
-        Pair<Optional<InstallmentNoPII>,Boolean> installmentAndPrimaryOrgFound = primaryOrgInstallmentPaidVerifierService.findAndValidatePrimaryOrgInstallment(primaryOrg, receiptDTO.getNoticeNumber());
-        primaryOrgFound[0] = installmentAndPrimaryOrgFound.getRight();
-        installmentAndPrimaryOrgFound.getLeft().ifPresent(installment -> {
-            log.debug("primaryOrg installment found id[{}]", installment.getInstallmentId());
-            Broker primaryBroker = brokerService.findById(primaryOrg.getBrokerId(), accessToken);
-            //update installment status
-            DebtPosition debtPosition = installmentUpdateService.updateInstallmentStatusOfDebtPosition(installment, primaryBroker, receiptDTO);
-            //persist updated debt position
-          //TODO
-            //start debt position workflow
-            DebtPositionDTO debtPositionDTO = debtPositionMapper.mapToDto(debtPosition);
-            invokeWorkflow(debtPositionDTO, accessToken);
-          });
-      });
+    //check if organization who handle the notice is managed by PU and update the installment status
+    boolean primaryOrgFound = updateInstallmentStatusOfDebtPosition.handleReceiptReceived(receiptDTO, accessToken);
 
     //for every organization handled by PU and mentioned in the receipt
-    createPaidTechnicalDebtPositionsService.createPaidTechnicalDebtPositionsFromReceipt(receiptDTO, !primaryOrgFound[0], accessToken);
+    createPaidTechnicalDebtPositionsService.createPaidTechnicalDebtPositionsFromReceipt(receiptDTO, !primaryOrgFound, accessToken);
 
     return receiptDTO;
   }
@@ -117,16 +76,6 @@ public class CreateReceiptServiceImpl implements CreateReceiptService {
   }
 
 
-  private void invokeWorkflow(DebtPositionDTO debtPositionDTO, String accessToken) {
-    if (!DebtPositionStatus.DRAFT.equals(debtPositionDTO.getStatus())) {
-      log.info("Invoking alignment workflow for debt position with id {}", debtPositionDTO.getDebtPositionId());
-      WorkflowCreatedDTO workflow = debtPositionSyncService.syncDebtPosition(debtPositionDTO, false, PaymentEventType.RT_RECEIVED, accessToken);
-      if (workflow != null) {
-        log.info("Workflow creation OK for debtPositionId[{}}: workflowId[{}]", debtPositionDTO.getDebtPositionId(), workflow.getWorkflowId());
-      } else {
-        log.warn("Workflow creation KO for debtPositionId[{}]: received null response", debtPositionDTO.getDebtPositionId());
-      }
-    }
-  }
+
 
 }
